+++ shared-trad-customer-serv/shared-trad-customer-impl/src/main/java/cn/suparking/customer/controller/order/service/OrderService.java
@@ -1,9 +1,9 @@
-package cn.suparking.customer.controller.parkOrder.service;
+package cn.suparking.customer.controller.order.service;
 
 import cn.suparking.common.api.beans.SpkCommonResult;
 import cn.suparking.order.api.beans.ParkingOrderQueryDTO;
 
-public interface ParkOrderService {
+public interface OrderService {
 
     /**
      * 根据用户ID,获取用户信息.
