+++ shared-trad-order-serv/shared-trad-order-impl/src/main/resources/mappers/car-group-order-sqlmap.xml
@@ -466,5 +466,4 @@
         WHERE
         id = #{id,jdbcType=BIGINT} and user_id = #{userId,jdbcType=BIGINT}
     </update>
-
 </mapper>
